requirements.txt
absl-py==2.1.0
attrs==24.2.0
cffi==1.17.1
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
editdistpy==0.1.5
flatbuffers==24.3.25
fonttools==4.54.1
jax==0.4.34
jaxlib==0.4.34
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
mediapipe==0.10.15
ml_dtypes==0.5.0
nltk==3.9.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.5
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.9.0.post0
regex==2024.9.11
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sounddevice==0.5.0
symspellpy==6.7.8
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.66.5
# Python Python 3.10.7
